% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outlier.R
\name{plot_outlier}
\alias{plot_outlier}
\title{Plot time series outliers / anomalies}
\usage{
plot_outlier(
  data,
  value_col,
  id_col = "ticker",
  date_col = "date",
  outlier_col = "Outlier",
  dot_size = 1,
  arrow_size = c(0.2, 0.8),
  ncol = 2
)
}
\arguments{
\item{data}{data.table object}

\item{value_col}{value column}

\item{id_col}{id column}

\item{date_col}{date column}

\item{outlier_col}{Outlier/Anomaly column, generated by \code{detect_outlier()}}

\item{dot_size}{size of dots in scatterplot}

\item{arrow_size}{vector of length 2 for arrow head and -line}

\item{ncol}{how much columns of plots? Default of 2}
}
\value{
cowplot object
}
\description{
Plot time series outliers / anomalies
}
\examples{
\dontrun{
  outlier <- detect_outlier(fcf::dow30[ticker == "AAPL",], value_col = "ebit")
  plot_outlier(outlier, value_col = "ebit")
}
}
