% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Quick statistical overview of your data}
\usage{
summary_statistics(data, cols)
}
\arguments{
\item{data}{data.frame}

\item{cols}{vector of column names to generate statistics for}
}
\value{
data.frame object with mean, mean, sd, min, max and N
}
\description{
Quick statistical overview of your data
}
